Move request handlers into a separate module
Split the implementation across multiple source files in response
to review feedback.

The `OPRF*` and `Error` types are directly-related to the implementation
of these particular handlers, so I kept those together. They need to
be public so the main application can call them, but otherwise are
just interfacing with the axum::route machinery.
